Separando config do Swagger em outra classe (arquivo).
