+++ src/main/kotlin/br/com/fiap/techfood/adapters/configs/BeanConfiguration.kt
@@ -40,23 +40,3 @@ class BeanConfiguration {
     }
 
 }
-
-@Configuration
-class SwaggerConfiguration {
-    @Bean
-    fun springShopOpenAPI(): OpenAPI {
-        return OpenAPI()
-            .info(
-                Info()
-                    .title("Tech Food")
-                    .description("Tech Food backend application")
-            )
-        // Vou deixar para referencia futura:
-        //.externalDocs(
-        //    ExternalDocumentation()
-        //        .description("SpringShop Wiki Documentation")
-        //        .url("https://springshop.wiki.github.org/docs")
-        //)
-    }
-}
-
